Files! Documentation, form control
Tweaked Gulp a bit, now VS Code debugger works properly for the server
and can place breakpoints (There was a source map issue)

Added a RAML file to describe the API (RESTful API Markup Language). Incomplete,
needs work

Refactored Form to SimpleForm (and RequestForm to SimpleRequestForm), going
to make another version which doesn't handle the layout at all in case
the developer wants to create a different layout than the one used (such
as for the search form)

Reworked FileInput to use the new FileDialogue

Changed the Input interface to remove parameters for onChange that weren't
needed, weren't used, and weren't implemented

Changed TextInput so that you could control the value being inputted.
Return false in the onChange parameter and it will not update the text
field

Added an image icon to be used with TextArea and FileDialogue to embed images

Added an API call to create folders

Fixed delete file bug where it wasn't referencing the correct file ID

FileInfo returns a complete FileObject. It used to return raw data, and
not have fields such as the ID or AccountID. Because it returned raw data,
memberOnly, forSlideshow, and forDisplay were numbers (0 or 1) instead of
booleans

Finished the PrettySQL tag function - didn't realize it wasn't working

Redid the FileDialogue, simplifying the process so that the all files were
being handled by one component, instead of one component controlling one
component per file and telling each to start uploading. The new process
just uploads the first file in its list and deletes it when done until
there are none left, passing the result from the upload (a FileObject)
to the controlling component. In theory, if there are new files added and
they aren't the same as the ones before, they will be added to the upload
queue (hasn't been tested)