More work with members
Continuing work with Members, nearly done with the NHQMember.
It can authenticate with NHQ, and get some database information. Everything
compiles and the NHQMember class can create a basic MemberObject with
permissions, cookies, and sessions

Created something like Linux PAM (Pluggable Authentication Modules) where there are
the code to interface with NHQ to get contacts and such is all modularized and can
in theory be easily replaced

Did refactoring to replace references to Member.ts with NHQMember.ts