Member state, standardization
Member state now works on the client - there were some issues where the wrong sessions
were being deleted.

Standardized the registry. It used to be a simple POST to get, now it is a GET to get
and has the framework in place to allow for POSTing, PATCHing, PUTing, etc

Added a check function to check if a session ID is valid. If so, it returns the associated
member as well to initialize the client app state