Tests! Member changes, CAPWATCH, classes
Added a gulp task for importing/exporting mysql database

Tests:
Changed test configuration to use a different database
Added an api route (/api/accountcheck) which echos the current account back
Added tests for:
        API 404
        Launching the server under different configurations
        Teapot
        Echo
        Member sessionss
        Accounts
        Blog posts
        Permissions
        Basic NHQMember functions
        All PAM modules for NHQ

Changed /api/check to uses enums instead of integers and strings

Changed /api/signin to use async/await and have better error handling

Refactored blog post apis to use the BlogPost class
Refactored some file apis to use the File class

Changed a couple of programs to stream JSON with async generators. As the async generator generates another item, it sends it as part of valid JSON

A lot of files have BaseMember replaced with MemberBase; the class was called MemberBase but the file was called BaseMember.ts

Added a function in Account to allow for generating a list of members that belong to an account

Created a BlogPost class; very similar to the Event class, wraps the BlogPostObject interface
Created a File class, similar to BlogPost class

Added a copy function to the event class. It copies all the data and moves times around given a delta. Options to copy status and files

Started work on importing CAPWATCH files. Structure is more modularized, now there are an array of modules which will each do their own thing
This function is an async generator, allowing for the caller to get results for each import separately

Added some datetime functions to the MySQL utils

Added streamAsyncGeneratorAsJSON and getTestTools utilities
streamAsyncGeneratorAsJSON will stream an async generator as JSON to a response object
getTestTools is used in unit tests

Added a CAPWATCHMember class which extends MemberBase to interface with the database to provide member information but to also differentiate from NHQMember

Added ORGID field to Members