Configuration, cleanup, efficiencies, typings
Configured a build task for VS Code, simplifies the startup process when trying
to programm

Added an API call to get the slideshow file IDs, as that is necessary at some
point

Created a new MySQL util which functions similar to collectResults. It's called
generateResults, and it returns an AsyncIterableIterator that will generate
rows. This allows programs to handle one row at a time and works well with
streamAsyncIteratorAsJSON

Renamed streamAsyncIteratorAsJSON to streamAsyncIteratorAsJSONArray

From here on, anything to deal with more than just a few items in an array
dealing with the database has to use generateResults and
streamAsyncIteratorAsJSONArray

Went through and cleaned up most API calls. This means I changed the format to
check for non-existance and throw an error, vs. checking for existance and
throwing an error in an else statement. Also went and made sure each error code
was appropriate, e.g. 403 for not signed in. Most APIs still need to check for
the member being signed in, and don't. Cleanup also included removing checking
for if req.account is undefined or null, as if that were the case it would be
handled by the Account.ExpressMiddleware function before each API is called
anyway

Events now work, and can be added/listed. Most of the work was done on the client

Created a new interface, DatabaseInterface<T>. This basically says that there
should be a method called toRaw() which returns T; will add other helpful
methods later (delete, save, set)

Files now check for the www accountID, this is optional in File.Get and should
be optional in methods that call File.Get

CAPWATCH imports now use async loops, which allows for them to not block the rest
of the server with their large for loops.

Fixed an issue with streamAsyncIteratorAsJSONArray where it may not provide the
first `[` in the JSON result if there is nothing to provide

TODO: As most APIs are being cleaned up, most of the business transactions
with the database are being moved to various classes in the lib folder.
Currently, there is nothing like that for creating folders, and
src/api/files/createfolder.ts interacts with the database directly