Did more work on events and attendance. Now, there are more functions than just creating and listing events, there are all the CRUD operations for attendance, CRUD operations for events, listing of events, and linking events
Su also works, there is no client UI for it however

Prospective members! There were 5 main methods created for it:
        1. Create - Creates a member in our database that can sign in and such
        2. Signin - Takes credentials and validates them against our database
        3. Get - Pulls the data from the database, like CAPWATCHMember
        4. Su - Used by NHQMember.su, to allow for developping
        5. ExpressMiddleware - Used by NHQMember.ExpressMiddleware to validate
                ProspectiveMember sessions. Caught by ExpressMiddleware vs.
                ConditionalExpressMiddleware

There is now a difference between NHQMember.ConditionalExpressMiddleware and
NHQMember.ExpressMiddleware. NHQMember.ExpressMiddleware works just like the
other, but if req.member ends up being null it terminates the connection with
a 403, never executing an API endpoint

NHQMember.ExpressMiddleware/ConditionalExpressMiddleware both support
ProspectiveMember, and therefore req.member now is the type `ProspectiveMember
 | NHQMember`.

Added attendance handling to Event.ts

Fixed a bug in streamAsyncGeneratorAsJSONArray where it would fail to stream
arrays longer than 1 because of a ternary operator ambiguity

Fixed a bug with generateResults as it would not generate fail to generate an
empty set if there were no items to generate

Fixed a bug with Event.Get, as typeof 1 !== typeof '1'. Types are coerced to
number now