Base drive, calendar, needed changes
Fixed circular dependency issues with members

Finished the File API for the most part, it is enough that
the client is able to have a good Drive

New APi call for getting a section of events by time for the calendar

NEEDED CHANGES IN DATATYPES:

Scrap member interfaces and create a better system. Currently, it is like this:

            MemberObject
                |
            MemberBase
        /       |           \
NHQMember   CAPWATCHMember  ProspectiveMemberAccount
                                        |
                                ProspectiveMember

It needs to be like this:

                              BaseMemberObject
                    /               |               \
NHQMemberObject             CAPWATCHMemberObject        ProspectiveMemberObject
    |   |                   /       |                       |           |
    |   NHQMember           |   CAPWATCHMember              |   Prospective...
    \-----------------------|-------------------------------/
                            |
                        MemberObject

Where MemberObject is a union type of CAPWATCHMemberObject, NHQMemberObject,
and ProspectiveMemberObject and the classes inherit from MemberBase as usual

This should resolve issues with conflicting ID types (string for Prosp. Member,
number for NHQ/CAPWATCH)

This will also scale nicely for if we introduce JROTC, AHG, BSA, etc.
Just add another class and interface pair

Will need to move ExpressMiddleware into MemberBase