Drive, calendar, button/form changes, email list
There is now an almost full drive, it just needs permission management!
A single file that, in its first commit, is almost 1200 lines long
Will need to abstract some components out, especially file uploader
as it could be used in FileDialogue

The calendar is probably done, there could be problems with events
overlapping but the algorithm MAY be able to handle it

Button and request forms now allow for selecting the HTTP method to use
Not as important for forms (can't really do DELETE or GET), but buttons
can now do DELETE and PUT alongside POST

There are 3 new form inputs, and one removed:

TextArea is technically being removed due to its size when included. Use
    LoadingTextArea instead, it's the same thing just a bit more space
    efficient for the client

BigTextBox is the old TextArea, it's just a big text input for when you
don't want a full WYSIWG editor but a TextInput is too small

DisabledMappedTextArea is just a display, a disabled text input

Code currently won't compile, need to fix that as I'm lazy

Email list is another page added, and it is to replace the email selector
page. It serves quite well, allowing for multiple filters to be used

I'm going to carry through, at some point, actually making Event, Member,
BlogPost, etc. classes for the client which will have some of the same
method names as their server-side counterparts but will use HTTP requests
to the correct server API to allow for some similar code to be used on the
client and the server. Should make some code easier to read, and make it
easier to manage fetch calls. Also, this will allow for easier unit testing,
as the classes can be mocked

Speaking of which, I will need to start unit tests for the client. I will
do so once I have those helper classes, which I will start after finishing
POCInputs, data types, and event viewer