Member refactor
Working on type refactor. Now the server code is done, there is just a
single compile time error in src/api/check.ts which TypeScript is happy
to compile and which provides no run time errors, so is it really an
error?

Most of the functionality of the code has not changed, only the
expression of types