Identified bug and fixed positioning bug
There is a bug with Chromium just for Linux that causes an element to be
lost when it is dragged and its parent is resized. It is identified to
only happen on Chromium for Linux; Chrome for Windows, Firefox, and
Microsoft Edge are all fine.

There was another bug identified for Firefox and Chrome for Windows
where when an element was dragged it was positioned 100px or so above
the cursor, causing confusion
This bug was fixed