Fixed tests, added SessionType, added CAPWATCH imports
About 10 tests were buggy and failing. There is still a flaky test, but
all the others are fixed

SessionType was added to Sessions. Represents regular, password reset,
and scan add sessions, each with special properties and limited access
to various bits of the website.

Added a simple CAPWATCH import API. It will probably be replaced later
with Server Sent Events or with a Websocket connection. Probably
websocket connection due to the support on the browser and the relative
complexity.