refactor: refactored linked events and attendance to use true links for events
Events are either links or event objects, not both. Attendance is stored at the central event
object. All views of an event for a linked event will pull the target event information and merge
the two. Attendance will pull the target event, but will filter based off of the account. No one has
permissions to modify the links themselves, nor the data of the central event. Permissions for
attendance are based off of the local event and check to make sure the record is viewable from the
local account perspective.

re #11