feat: added the events and personalfolders folders
affects: client, evmplus-org-server, server-common

The Events folder represents where files which are to be attached to events are uploaded by default.
This folder is available to everyone who has the ability to manage events.

There is now a personalfolders folder which contains all of the personal folders of everyone in the
unit, unless that folder is empty. This folder has no permissions, making it accessible only to
admins who have site permissions that override the Drive permissions